requirements.txt
requests
rich
click
python-dotenv
boto3
pymupdf
pillow